packaging: Full qualified imports plus eta_script.py, a script that allows running ETA from the directory.
We want to use ETA in two ways. Installed as a module, used for normal user installation but also in future automatic tests. For
local development it is also useful to launch ETA from a folder (e.g. comparing two versions). For this we added the eta_script.py file, that messes with your sys.path.

Fixes #118.