logging: Move any logoutput to the python logging facilities. To communicate with the frontend the logger ´etabackend.frontend´ (See ETA_CUT.__init__()) is used.
An additional handler is configured to send the logstream on ´etabackend.frontend´ to the frontend. There is still some status update messages using the verbose() function necessary.
These are marked with FIXME for the future.
The current state should reflect the same state as before the switch to logging.

This commit relates to #114.