backend: Runs bokeh server on main asyncio loop. Changes eta.display() function. Enables Bokeh support again.
The bokeh server is now run on the main asyncio loop, instead of the ThreadPoolWorker. This is achieved through queuing backend.show_bokeh using the asyncio.loop with a function creating the bokeh document.
It is important that this functions returns fast and is not used to do calculations within the eta evaluation logic. Calculate everything before.

The backend.display function is changed to contain an argument on the dashboard system, the user wants to use. With the default currently still being dash to support plotly. If you want to use bokeh, the type parameter should be 'bokeh'.